TEMPORARY: Please read commit message
cosmic-applets-power needs a new version of zbus, as this crate
uses zbus, this conflict needed to be resolved by using the main
branch in this repo too.

Though this will not be necessary once zbus 3.6.3 is released.
After 3.6.3 is released please update the cosmic-applets to point
to the release version not this repo, and then delete this repo.